Remove LDN processor and test
Migrating the LDNProcessor to fcrepo5 is fraught, especially when
considering the prospect of supporting both fcrepo4-style messages, and
fcrepo5-style.  Additionally, this processor is not used in
fcrepo-camel-toolbox.  Therefore, removing it is the most
straightforward path forward.  If a need is discovered for it, it can be
reinstated, with a little thought.